Move `components.py` to `/app/runner/`
